Alliance of Consumer Complaints Using NLP
Here we have a complaint data that I conclude by using web scrapping, it was really an osm experience to have it, let's talk about data there are two features, those are "Narrative " and "Product", in this project I just clean the words and based on that conclude the model, there I use some of the new model matrixes. also, I a new model that I used for the first time.